XIVY-13128 Cleanup prototype - Reduce any's - Restructure files
