+++ packages/editor/src/components/blocks/text/Text.css
