--- packages/editor/src/components/data/data.ts
@@ -1,13 +1,0 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-import type { UiComponentProps } from '../components/component';
-
-export type ContentData = {
-  id: string;
-  type: string;
-  props: UiComponentProps<any>;
-};
-
-export type UiEditorData = {
-  root: any;
-  content: ContentData[];
-};
