+++ packages/editor/src/components/editor/properties/PropertyItem.tsx
@@ -1,12 +1,12 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
 import { useEffect, useState } from 'react';
-import type { Field } from '../../components/component';
-import { useData } from '../../data/useData';
+import type { Field, PrimitiveValue } from '../../../types/config';
+import { useData } from '../../../data/useData';
 import './PropertyItem.css';
 import { InputField } from './fields/InputField';
 import { CheckboxField } from './fields/CheckboxField';
 import { TextareaField } from './fields/TextareaField';
 import { SelectField } from './fields/SelectField';
+import { NumberField } from './fields/NumberField';
 
 type PropertyItemProps = {
   fieldName: string;
@@ -15,8 +15,8 @@ type PropertyItemProps = {
 
 export const PropertyItem = ({ fieldName, field }: PropertyItemProps) => {
   const { element, setElement } = useData();
-  const [value, setValue] = useState<any>();
-  const onChange = (newValue: any) => {
+  const [value, setValue] = useState<PrimitiveValue>();
+  const onChange = (newValue: PrimitiveValue) => {
     setValue(newValue);
     if (element) {
       element.props[fieldName] = newValue;
@@ -29,12 +29,13 @@ export const PropertyItem = ({ fieldName, field }: PropertyItemProps) => {
   const inputFor = (field: Field) => {
     switch (field.type) {
       case 'text':
+        return <InputField field={field} value={value} onChange={onChange} />;
       case 'number':
-        return <InputField field={field} value={value ?? ''} onChange={onChange} />;
+        return <NumberField field={field} value={value} onChange={onChange} />;
       case 'checkbox':
-        return <CheckboxField field={field} value={value ?? false} onChange={onChange} />;
+        return <CheckboxField field={field} value={value} onChange={onChange} />;
       case 'textarea':
-        return <TextareaField value={value ?? ''} onChange={onChange} />;
+        return <TextareaField value={value} onChange={onChange} />;
       case 'select':
         return <SelectField field={field} value={value} onChange={onChange} />;
       default:
