+++ packages/editor/src/components/editor/properties/fields/CheckboxField.tsx
@@ -1,11 +1,11 @@
-import type { Field } from '../../../components/component';
+import type { Field, PrimitiveValue } from '../../../../types/config';
 
 type CheckboxFieldProps = {
   field: Field;
-  value: boolean;
+  value?: PrimitiveValue;
   onChange: (value: boolean) => void;
 };
 
 export const CheckboxField = ({ field, value, onChange }: CheckboxFieldProps) => (
-  <input type={field.type} checked={value} onChange={e => onChange(e.target.checked)} />
+  <input type={field.type} checked={(value ?? false) as boolean} onChange={e => onChange(e.target.checked)} />
 );
