+++ packages/editor/src/components/editor/properties/fields/InputField.tsx
@@ -1,12 +1,11 @@
-import type { Field } from '../../../components/component';
+import type { Field, PrimitiveValue } from '../../../../types/config';
 
 type InputFieldProps = {
   field: Field;
-  value: string | number;
+  value?: PrimitiveValue;
   onChange: (value: string | number) => void;
 };
 
 export const InputField = ({ field, value, onChange }: InputFieldProps) => {
-  const onInternalChange = (value: string) => (field.type === 'number' ? onChange(Number(value)) : onChange(value));
-  return <input type={field.type} value={value} onChange={e => onInternalChange(e.target.value)} />;
+  return <input type={field.type} value={(value ?? '') as string} onChange={e => onChange(e.target.value)} />;
 };
