+++ packages/editor/src/data/data.ts
@@ -1,0 +1,12 @@
+import type { DefaultComponentProps } from '../types/config';
+
+export type ContentData = {
+  id: string;
+  type: string;
+  props: DefaultComponentProps;
+};
+
+export type UiEditorData = {
+  root: {};
+  content: ContentData[];
+};
