+++ packages/editor/src/utils/array.test.ts
